pbd/pipelines/pretrain/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.2.1
colorama==0.4.6 ; sys_platform == 'win32'
datasets==4.4.1
dill==0.4.0
filelock==3.20.0
frozenlist==1.8.0
fsspec==2025.10.0
gitdb==4.0.12
gitpython==3.1.45
h11==0.16.0
hf-xet==1.2.0 ; platform_machine == 'aarch64' or platform_machine == 'amd64' or platform_machine == 'arm64' or platform_machine == 'x86_64'
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
multidict==6.7.0
multiprocess==0.70.18
numpy==2.3.5
packaging==25.0
pandas==2.3.3
platformdirs==4.5.0
propcache==0.4.1
protobuf==5.29.5
pyarrow==22.0.0
pydantic==2.11.9
pydantic-core==2.33.2
python-dateutil==2.9.0.post0
pytz==2025.2
pyyaml==6.0.3
regex==2025.11.3
requests==2.32.5
safetensors==0.7.0
sentry-sdk==2.46.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
tokenizers==0.22.1
tqdm==4.67.1
transformers==4.57.3
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.2
urllib3==2.3.0
wandb==0.23.0
xxhash==3.6.0
yarl==1.22.0
